processing/requirements.txt
apache-flink==1.17.0
# wasmtime-py
# wasmer-python
kafka-python==2.0.2
numpy==1.24.3
scipy==1.10.1
pandas==2.0.2
librosa==0.10.1
opencv-python==4.8.0.76
